+++ lldb/source/Plugins/Process/Utility/DynamicRegisterInfo.cpp
@@ -632,10 +632,10 @@ DynamicRegisterInfo::Dump () const
 {
     StreamFile s(stdout, false);
     const size_t num_regs = m_regs.size();
-    s.Printf("%p: DynamicRegisterInfo contains %zu registers:\n", this, num_regs);
+    s.Printf("%p: DynamicRegisterInfo contains %" PRIu64 " registers:\n", this, (uint64_t)num_regs);
     for (size_t i=0; i<num_regs; ++i)
     {
-        s.Printf("[%3zu] name = %-10s", i, m_regs[i].name);
+        s.Printf("[%3" PRId64 "] name = %-10s", i, m_regs[i].name);
         s.Printf(", size = %2u, offset = %4u, encoding = %u, format = %-10s",
                  m_regs[i].byte_size,
                  m_regs[i].byte_offset,
@@ -673,10 +673,10 @@ DynamicRegisterInfo::Dump () const
     }
     
     const size_t num_sets = m_sets.size();
-    s.Printf("%p: DynamicRegisterInfo contains %zu register sets:\n", this, num_sets);
+    s.Printf("%p: DynamicRegisterInfo contains %" PRIu64 " register sets:\n", this, (uint64_t)num_sets);
     for (size_t i=0; i<num_sets; ++i)
     {
-        s.Printf("set[%zu] name = %s, regs = [", i, m_sets[i].name);
+        s.Printf("set[%" PRIu64 "] name = %s, regs = [", (uint64_t)i, m_sets[i].name);
         for (size_t idx=0; idx<m_sets[i].num_registers; ++idx)
         {
             s.Printf("%s ", m_regs[m_sets[i].registers[idx]].name);
